ダメージ判定の実装など
・プレイヤーのHP表示を管理するHeartManagerとHeartImageControllerの追加
・触れると敵が動き出す、スポーンエリアをカメラにつけました。

+++ Assets/Yoshioka/Scripts/IDamagable.cs
@@ -1,4 +1,4 @@
 public interface IDamagable
 {
-    void AddDamage(float damage);
+    void AddDamage(int damage);
 }
